refactor: useStudyRecord 훅 분리
